refactor: extract log modal viewer from flow-view into dedicated module
